pyproject.toml
[project]
name = "readykit"
version = "1.4.0"
description = "Production-ready Flask SaaS template with multi-tenant workspaces, Stripe billing, and OAuth"
authors = [
    {name = "Level09", email = "hello@level09.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"

dependencies = [
    # Core Flask and web framework
    "flask>=3.1.1",
    "werkzeug>=3.1.3",
    "jinja2>=3.1.6",
    "click>=8.2.1",
    "itsdangerous>=2.2.0",
    "markupsafe>=3.0.2",
    "blinker>=1.9.0",
    # Flask extensions (production)
    "flask-sqlalchemy>=3.1.1",
    "flask-security-too>=5.6.2",
    "flask-session>=0.8.0",
    "flask-caching>=2.3.1",
    "flask-mail>=0.10.0",
    "flask-wtf>=1.2.2",
    "flask-babel>=4.0.0",
    "flask-login>=0.6.3",
    "flask-dance>=7.1.0",
    # Database and ORM
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.10",
    # Authentication and security
    "argon2-cffi>=25.1.0",
    "passlib[argon2]>=1.7.4",
    "bcrypt>=4.3.0",
    "cryptography>=45.0.5",
    "webauthn>=2.6.0",
    "cffi>=1.17.1",
    "pycparser>=2.22",
    # Utilities and data handling
    "email-validator>=2.2.0",
    "python-dotenv>=1.1.1",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "orjson>=3.10.18",
    "rich>=14.0.0",
    "qrcode>=8.2",
    "pillow>=11.3.0",
    "bleach>=6.2.0",
    # Legacy/compatibility
    "flask-script>=2.0.6",
    "babel>=2.17.0",
    "speaklater>=1.3",
    "six>=1.17.0",
    "wtforms>=3.2.1",
    "setuptools>=80.9.0",
    "mako>=1.3.10",
    "python-editor>=1.0.4",
    # Production server
    "uwsgi>=2.0.30; sys_platform != 'win32'",
    "gunicorn>=23.0.0",
    "sluggi>=0.1.1",
    # Payments and billing
    "stripe>=12.0.0",
]

[project.optional-dependencies]
full = [
    "redis>=7.1.0",
    "celery>=5.5.3",
    "kombu>=5.5.4",
    "amqp>=5.3.1",
    "vine>=5.1.0",
]
dev = [
    # Code formatting and linting
    "ruff",
    "pre-commit",

    # Development tools
    "flask-debugtoolbar>=0.16.0",
    "flask-shell-ipython>=0.5.3",
]
wsgi = [
    "uwsgi>=2.0.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["enferno"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",   # line too long, handled by formatter
    "B008",   # do not perform function calls in argument defaults
    "C901",   # complexity
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["enferno", "readykit"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "flask-shell-ipython>=0.5.3",
]
